Move function Ioctl into kvm/ioctl.go
